requirements.txt
google-genai
pyaudio
numpy
matplotlib
gtts